+++ lib/bh/railtie.rb
@@ -1,4 +1,5 @@
 require 'bh/helpers/alert_helper'
+require 'bh/helpers/button_helper'
 require 'bh/helpers/button_to_helper'
 require 'bh/helpers/cdn_helper'
 require 'bh/helpers/dropdown_helper'
@@ -17,6 +18,7 @@ module Bh
   class Railtie < Rails::Railtie
     initializer 'bh.add_helpers' do
       ActionView::Base.send :include, AlertHelper
+      ActionView::Base.send :include, ButtonHelper
       ActionView::Base.send :include, ButtonToHelper
       ActionView::Base.send :include, CdnHelper
       ActionView::Base.send :include, DropdownHelper